refs #67 filterをpackageにまとめる
