+++ server.go
@@ -2,14 +2,12 @@ package main
 import (
 	"flag"
 	"net/http"
-	"fmt"
-	"syscall"
-	"strconv"
 	"github.com/zenazn/goji"
 	"github.com/zenazn/goji/web"
 	"github.com/flosch/pongo2"
 	_ "github.com/flosch/pongo2-addons"
 	"./controllers"
+	"./filters"
 )
 
 
@@ -32,23 +30,8 @@ func Routes(m *web.Mux) {
 }
 
 func main() {
-	pongo2.RegisterFilter("suffixAssetsUpdate", SuffixAssetsUpdate)
+	pongo2.RegisterFilter("suffixAssetsUpdate", filters.SuffixAssetsUpdate)
 	flag.Set("bind", ":9090")
 	Routes(goji.DefaultMux)
 	goji.Serve()
-}
-
-func SuffixAssetsUpdate(in *pongo2.Value, param *pongo2.Value) (out *pongo2.Value, err *pongo2.Error) {
-	assetsFile, ok := in.Interface().(string)
-	if !ok {
-		return nil, &pongo2.Error{
-			Sender: "suffixStylesheet",
-			ErrorMsg: fmt.Sprintf("Data must be string %T ('%v')", in, in),
-		}
-	}
-
-	var file syscall.Stat_t
-	syscall.Stat("./public/assets" + assetsFile, &file)
-	timestamp, _ := file.Mtim.Unix()
-	return pongo2.AsValue(assetsFile + "?update=" + strconv.FormatInt(timestamp, 10)), nil
 }
